Update README.md
The Air Traffic Analysis Project is an in-depth exploration and visualization of air traffic data from domestic and international flights in the United States. Utilizing a dataset sourced from Kaggle, this project aims to uncover key trends, patterns, and insights related to air travel. The analysis leverages Excel for data manipulation, visualization, and forecasting, providing a comprehensive dashboard and detailed answers to critical research questions.